requirements-nodjango.txt
Pillow==2.3.0
psycopg2==2.6
Unipath==1.0
beautifulsoup4==4.3.2
boto==2.35.2
coverage==3.7.1
colour_runner==0.0.4
django-annoying==0.8.1
django-bootstrap3==4.8.2
django-braces==1.3.1
django-coverage==1.2.4
django-crispy-forms==1.4.0
django-dbbackup==2.0.4
django-debug-toolbar==1.3.0
django-debug-toolbar-template-timings==0.6.4
django-fixture-media==0.1.4
django-floppyforms==1.1.1
# django-model-utils==2.2
# because newer version that fixes migrations on 1.7 is not released yet
git+https://github.com/carljm/django-model-utils.git@d797996d13a871328bada34aef8ad25534e0058f
django-rosetta==0.7.3
#django-sitemetrics==0.4.2
git+https://github.com/idlesign/django-sitemetrics.git@886b35443d7163b8cab7a49c57bd7be6337379ef
django-taggit==0.14.0
factory_boy==2.5.2
icalendar==3.8.4
mock==1.0.1
newrelic==2.44.0.36
pylint-django==0.3
python-dateutil==2.2
pytz==2014.10
raven==5.0.0
requests==2.2.0
six==1.5.2
slideshare==0.1dev
sorl-thumbnail==12.2
sqlparse==0.1.11
testfixtures==3.0.2
yandexwebdav==0.2.11
django-pipeline==1.5.1
unicodecsv==0.13.0
